python project
This project aims to application the KGB game using artificial 
intelligence algorithms. In this project, we have implemented a 
stochastic algorithm that allows computer to look many moves 
ahead, thus picking the move that optimizes its chances of the 
winning.The draw is done randomly for each of the players, and the choice will 
fall on (girl / boy / king). The king can beat the boy, the boy can beat 
the girl, but only the girl can beat the king, and the winner is achieved 
based on this method.